% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looker_api_functions.R
\name{get_looker_token}
\alias{get_looker_token}
\title{get_looker_token}
\usage{
get_looker_token(client, secret)
}
\arguments{
\item{client}{Looker API3 client, obtained by generating a new client using the user administation dashboard in the Looker GUI.
This should be stored in an environment variable and NEVER written in your code or shared.}

\item{secret}{Looker API3 client, obtained by generating a new client using the user administation dashboard in the Looker GUI.
This should be stored in an environment variable and NEVER written in your code or shared.}
}
\value{
A character string containing the session token. This can either be passed directly to
 \code{run_look()} or stored in an R object to be passed as an argument to \code{run_look()}.
 The latter is the preferred method if you are going to be running more than a single look.
}
\description{
get_looker_token
}
\examples{

get_looker_token(Sys.getenv("LOOKER_API_3_CLIENT"),
Sys.getenv("LOOKER_API_3_SECRET"))

}

