% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looker_api.R
\name{run_look}
\alias{run_look}
\title{run_look}
\usage{
run_look(look_id, token = NA, client = Sys.getenv("LOOKER_API_3_CLIENT"),
  secret = Sys.getenv("LOOKER_API_3_SECRET"),
  base_url = "https://upworthy.looker.com", limit = 10000, cache = TRUE)
}
\arguments{
\item{look_id}{The ID of a saved look in looker. This is the number at the end of the URL when you open a saved
look in the GUI}

\item{token}{The session token generated by \code{get_session_token}, if NA then a new token is generated}

\item{client}{Your looker API 3 client, remember to store this in an environment variable}

\item{secret}{Your looker API 3 client, remember to store this in an environment variable}

\item{base_url}{The looker URL for your company for Upworthy this is always "https://upworthy.looker.com"}

\item{limit}{Number of rows to return, will over-ride the settings in the saved look}

\item{cache}{Whether to return a cached value if available. Generally, you should use the default value}
}
\value{
A dataframe containing the result of running the look.
}
\description{
run_look
}
\examples{

# Replace this look id with one you have access to in the GUI.

run_look(1969)

#Save the token in an intermediary object if running multiple looks

token <- get_looker_token(Sys.getenv("LOOKER_API_3_CLIENT"),
Sys.getenv("LOOKER_API_3_SECRET"))

run_look(1969, token)
run_look(1972, token)

}

